refactor:拆分MemoryGenerationTask至单独编排组件
