+++ src/hivememory/patchouli/services/__init__.py
@@ -1,6 +1,7 @@
 """Patchouli internal microservices."""
 
 from hivememory.patchouli.services.librarian import LibrarianCore
+from hivememory.patchouli.services.memory_generation_tasks import MemoryGenerationTaskController
 from hivememory.patchouli.services.retrieval import RetrievalFamiliar
 
-__all__ = ["LibrarianCore", "RetrievalFamiliar"]
+__all__ = ["LibrarianCore", "MemoryGenerationTaskController", "RetrievalFamiliar"]
